get good camera position for curr field
