+++ src/frontend/lib/views/addfield/add_field_details_screen.dart
@@ -5,6 +5,7 @@ import 'package:frontend/views/addfield/widgets/add_field_info_card.dart';
 import 'package:frontend/views/addfield/widgets/visualize_field_map.dart';
 import 'package:frontend/widgets/terrafarm_app_bar.dart';
 import 'package:frontend/widgets/terrafarm_rounded_button.dart';
+import 'package:frontend/utils/polygon_utils.dart' as utils;
 import 'package:provider/provider.dart';
 
 class AddFieldDetailsScreen extends StatefulWidget {
@@ -54,8 +55,7 @@ class _AddFieldDetailsScreenState extends State<AddFieldDetailsScreen> {
                         // Gets the polygon to visualize from the provider.
                         polygon: Provider.of<NewFieldProvider>(context).getPolygon(),
                         // Gets a good camera position.
-                        cameraPosition: Provider.of<NewFieldProvider>(context)
-                            .getGoodCameraPositionForPolygon(),
+                        cameraPosition: utils.getGoodCameraPositionForPolygon(Provider.of<NewFieldProvider>(context).geoPoints),
                       ),
                     ),
                     SizedBox.fromSize(
